+++ data/src/main/java/com/yapp/android2/data/remote/login/LoginRemoteDataSourceImpl.kt
@@ -3,6 +3,8 @@ package com.yapp.android2.data.remote.login
 import com.yapp.android2.data.service.LoginService
 import com.yapp.android2.domain.entity.LoginRequest
 import com.yapp.android2.domain.entity.LoginResponse
+import com.yapp.android2.domain.entity.NotificationRequest
+import com.yapp.android2.domain.entity.NotificationResponse
 import javax.inject.Inject
 
 class LoginRemoteDataSourceImpl @Inject constructor(
@@ -13,4 +15,7 @@ class LoginRemoteDataSourceImpl @Inject constructor(
         return loginService.postLogin(request)
     }
 
+    override suspend fun postFCMToken(request: NotificationRequest): NotificationResponse {
+        return loginService.postFCMToken(request)
+    }
 }
