+++ data/src/main/java/com/yapp/android2/data/repository/LoginRepositoryImpl.kt
@@ -2,10 +2,7 @@ package com.yapp.android2.data.repository
 
 import com.yapp.android2.data.local.login.LoginLocalDataSource
 import com.yapp.android2.data.remote.login.LoginRemoteDataSource
-import com.yapp.android2.domain.entity.Login
-import com.yapp.android2.domain.entity.LoginRequest
-import com.yapp.android2.domain.entity.LoginResponse
-import com.yapp.android2.domain.entity.User
+import com.yapp.android2.domain.entity.*
 import com.yapp.android2.domain.repository.login.LoginRepository
 import javax.inject.Inject
 
@@ -16,6 +13,10 @@ class LoginRepositoryImpl @Inject constructor(
 
     override suspend fun postLogin(request: LoginRequest): LoginResponse {
         return loginRemoteDataSource.postLogin(request)
+    }
+
+    override suspend fun postFCMToken(request: NotificationRequest): NotificationResponse {
+        return loginRemoteDataSource.postFCMToken(request)
     }
 
     override fun saveAccessToken(token: String) {
