+++ data/src/main/java/com/yapp/android2/data/repository/NotificationRepositoryImpl.kt
@@ -1,8 +1,6 @@
 package com.yapp.android2.data.repository
 
 import com.yapp.android2.data.remote.notification.NotificationRemoteDataSource
-import com.yapp.android2.domain.entity.NotificationRequest
-import com.yapp.android2.domain.entity.NotificationResponse
 import com.yapp.android2.domain.repository.Notification.NotificationRepository
 import javax.inject.Inject
 
@@ -10,7 +8,4 @@ class NotificationRepositoryImpl @Inject constructor(
     private val notificationRemoteDataSource: NotificationRemoteDataSource
 ): NotificationRepository {
 
-    override suspend fun postFCMToken(request: NotificationRequest): NotificationResponse {
-        return notificationRemoteDataSource.postFCMToken(request)
-    }
 }
