+++ data/src/main/java/com/yapp/android2/data/service/LoginService.kt
@@ -2,6 +2,8 @@ package com.yapp.android2.data.service
 
 import com.yapp.android2.domain.entity.LoginRequest
 import com.yapp.android2.domain.entity.LoginResponse
+import com.yapp.android2.domain.entity.NotificationRequest
+import com.yapp.android2.domain.entity.NotificationResponse
 import retrofit2.http.Body
 import retrofit2.http.POST
 
@@ -12,4 +14,8 @@ interface LoginService : Service {
         @Body request: LoginRequest
     ): LoginResponse
 
+    @POST("/api/fcm-token")
+    suspend fun postFCMToken(
+        @Body request: NotificationRequest
+    ): NotificationResponse
 }
