+++ domain/src/main/java/com/yapp/android2/domain/repository/login/LoginRepository.kt
@@ -1,12 +1,11 @@
 package com.yapp.android2.domain.repository.login
 
-import com.yapp.android2.domain.entity.LoginRequest
-import com.yapp.android2.domain.entity.LoginResponse
-import com.yapp.android2.domain.entity.User
+import com.yapp.android2.domain.entity.*
 import com.yapp.android2.domain.repository.Repository
 
 interface LoginRepository : Repository {
     suspend fun postLogin(request: LoginRequest): LoginResponse
+    suspend fun postFCMToken(request: NotificationRequest): NotificationResponse
     fun saveAccessToken(token: String)
     fun getAccessToken(): String
     fun saveKakaoAccessToken(kakaoToken: String)
