+++ domain/src/main/java/com/yapp/android2/domain/usecase/LoginUseCase.kt
@@ -1,8 +1,6 @@
 package com.yapp.android2.domain.usecase
 
-import com.yapp.android2.domain.entity.LoginRequest
-import com.yapp.android2.domain.entity.LoginResponse
-import com.yapp.android2.domain.entity.User
+import com.yapp.android2.domain.entity.*
 import com.yapp.android2.domain.repository.login.LoginRepository
 import javax.inject.Inject
 
@@ -14,11 +12,15 @@ class LoginUseCase @Inject constructor(
         return loginRepository.postLogin(request)
     }
 
+    suspend fun postFCMToken(request: NotificationRequest): NotificationResponse {
+        return loginRepository.postFCMToken(request)
+    }
+
     fun saveAccessToken(token: String) {
         loginRepository.saveAccessToken(token)
     }
 
-    fun saveKakaoAccessToken(token: String){
+    fun saveKakaoAccessToken(token: String) {
         loginRepository.saveKakaoAccessToken(token)
     }
